Generate a Kotlin typealias to Any when type cannot be identified
This supports a JSON schema type that allows all valid JSON. We are
building an `Untyped` type that looks like the following:

```javascript
"Untyped": {
  "type": [
    "array",
    "boolean",
    "integer",
    "number",
    "object",
    "string"
  ]
}
```

Currently, Kotlin code is not generated for this type because a
single type cannot be identified.

This PR adds an allowance for this and generates a `typealias = Any`
to allow named references to this type.